Remove some aliases that can better be implemented with the new abbreviation system
Add (commented out) system for notifications when a command that
took longer than x seconds terminates. Currently due to a bug in fish
the last command line is not yet available when fish_prompt is executed.
It would be preferrable to only activate this behavior for certain
commands (e.g. commands that do not involve user interaction).
For this to work, the command must be known.
Alternative: define a function that appends a notification to a command
and that must explicitly be called